Run command for a target amount of time
This patch adds a command line parameter which allows the user to
specify the target amount of time to spend gathering results, and the
minimum number of iterations to perform.

Instead of executing the command for a fixed number of iterations, the
command is repeatedly executed until the target amount of time has been
spent, or the minimum number of results has been gathered.