Add a command line parameter to graph the results
If invoked with the -g or --graph flags set, srtime will show a scatter
plot of the execution times for all iterations after running. This can
be useful for detecting trends in the execution time that would
otherwise not be spotted, for example a trend towards lower execution
times over each iteration (likely a result of caching).