process: Add a "quiet" option to disable stdout
This modifies the behaviour of the Process class so that it will
suppress output to stdout if invoked with the "quiet" option.

Note that this patch does not implement a command line argument for the
quiet option. Since the srtime CLI prints output to stderr, the simplest
way to discard the output of a process is to simply pipe stdout, e.g.

   srtime ./benchmark >/dev/null